+++ R/EmpiricalCalibration.R
@@ -1,6 +1,6 @@
 # @file EmpiricalCalibration.R
 #
-# Copyright 2020 Observational Health Data Sciences and Informatics
+# Copyright 2021 Observational Health Data Sciences and Informatics
 #
 # This file is part of EmpiricalCalibration
 # 
@@ -23,4 +23,5 @@
 #' @importFrom grDevices rgb
 #' @importFrom utils read.csv
 #' @importFrom methods is
+#' @importFrom rlang .data
 NULL