+++ NEWS.md
@@ -1,3 +1,12 @@
+EmpiricalCalibration 2.1.0
+==========================
+
+Changes
+
+1. Adding computation of expected absolute systematic error. Can be shown in plots.
+
+
+
 EmpiricalCalibration 2.0.2
 ==========================
 
@@ -43,7 +52,7 @@ EmpiricalCalibration 1.4.0
 
 Changes
 
-1. Added plot showing effec of confidence interval calibration, similar to p-value plot.
+1. Added plot showing effect of confidence interval calibration, similar to p-value plot.
 
 Bugfixes